+++ buildsystem/spack/ascent/install.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+bsub ./buildsystem/spack/ascent/bsub.sh